requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.7.6
    # via -r requirements.in
appnope==0.1.2
    # via ipython
astroid==2.9.3
    # via pylint
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via
    #   jsonschema
    #   pytest
backcall==0.2.0
    # via ipython
black==22.1.0
    # via ipython
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.11
    # via requests
click==8.0.3
    # via
    #   black
    #   clickclick
    #   flask
clickclick==20.10.2
    # via connexion
connexion[swagger-ui]==2.11.0
    # via -r requirements.in
coverage[toml]==6.3.1
    # via pytest-cov
decorator==5.1.1
    # via ipython
dependency-injector==4.38.0
    # via -r requirements.in
dramatiq[rabbitmq]==1.12.3
    # via -r requirements.in
executing==0.8.2
    # via stack-data
factory-boy==3.2.1
    # via pytest-factoryboy
faker==12.1.0
    # via factory-boy
flask==2.0.2
    # via connexion
greenlet==1.1.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
inflection==0.5.1
    # via
    #   connexion
    #   pytest-factoryboy
iniconfig==1.1.1
    # via pytest
ipython==8.0.1
    # via -r requirements.in
isodate==0.6.1
    # via openapi-schema-validator
isort==5.10.1
    # via pylint
itsdangerous==2.0.1
    # via flask
jedi==0.18.1
    # via ipython
jinja2==3.0.3
    # via
    #   flask
    #   swagger-ui-bundle
jsonschema==3.2.0
    # via
    #   connexion
    #   openapi-schema-validator
    #   openapi-spec-validator
lazy-object-proxy==1.7.1
    # via astroid
mako==1.1.6
    # via alembic
markupsafe==2.0.1
    # via
    #   jinja2
    #   mako
matplotlib-inline==0.1.3
    # via ipython
mccabe==0.6.1
    # via pylint
mypy-extensions==0.4.3
    # via black
openapi-schema-validator==0.1.6
    # via openapi-spec-validator
openapi-spec-validator==0.3.3
    # via connexion
packaging==21.3
    # via pytest
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pdf2image==1.16.0
    # via -r requirements.in
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pika==1.2.0
    # via dramatiq
pillow==9.0.1
    # via pdf2image
platformdirs==2.4.1
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
prometheus-client==0.13.1
    # via dramatiq
prompt-toolkit==3.0.26
    # via ipython
psycopg2-binary==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via pytest
pydantic==1.9.0
    # via -r requirements.in
pygments==2.11.2
    # via ipython
pylint==2.12.2
    # via -r requirements.in
pyparsing==3.0.7
    # via packaging
pypdf2==1.26.0
    # via -r requirements.in
pyrsistent==0.16.1
    # via
    #   jsonschema
    #   openapi-spec-validator
pytest==7.0.0
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-factoryboy
    #   pytest-mock
pytest-cov==3.0.0
    # via -r requirements.in
pytest-factoryboy==2.1.0
    # via -r requirements.in
pytest-mock==3.7.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
pyyaml==5.4.1
    # via
    #   clickclick
    #   connexion
    #   openapi-spec-validator
requests==2.27.1
    # via
    #   connexion
    #   requests-toolbelt
requests-toolbelt==0.9.1
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   dependency-injector
    #   isodate
    #   jsonschema
    #   openapi-schema-validator
    #   openapi-spec-validator
    #   pyrsistent
    #   python-dateutil
sqlalchemy==1.4.31
    # via
    #   -r requirements.in
    #   alembic
stack-data==0.1.4
    # via ipython
structlog==21.5.0
    # via -r requirements.in
swagger-ui-bundle==0.0.9
    # via connexion
toml==0.10.2
    # via pylint
tomli==2.0.0
    # via
    #   black
    #   coverage
    #   pytest
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.0.1
    # via pydantic
urllib3==1.26.8
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.0.2
    # via
    #   connexion
    #   flask
wrapt==1.13.3
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
